pyproject.toml
[project]
name = "ai-agent-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai==1.12.1",
    "pexpect>=4.9.0",
    "pytest>=9.0.2",
    "python-dotenv==1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
# Exclude common directories
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "*.egg-info",
    "build",
    "dist",
]

# Set line length
line-length = 120

# Target Python 3.11+
target-version = "py311"

[tool.ruff.lint]
# Enable recommended rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules
ignore = [
    "E501",  # Line too long (handled by formatter)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Use Unix line endings
line-ending = "lf"

[tool.ruff.lint.isort]
# Group imports
known-first-party = ["functions", "logger", "config", "prompts", "call_function"]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]
